requirements.txt
pynetworktables
numpy==1.18.1
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
